move run_dungeon() to seperate file.
