+++ run_dungeon.h
@@ -1,0 +1,10 @@
+#ifndef RUN_DUNGEON_H
+#define RUN_DUNGEON_H
+
+/**
+ * run_dungeon() - dungeon code
+ * dungeon_num - Dungeon Number to run
+ */
+void run_dungeon(unsigned char dungeon_num);
+
+#endif /* RUN_DUNGEON_H */
